HW: change CC pull scheme to ensure delivery of Vbus from PixC
Fixed pulldown is routed through jumper footprints, with the pulldown for CC2
hard-shorted across the jumper. This allows moving the hard pulldown from CC2
to CC1 if testing reveals a mistake was made in selecting the correct CCn pin.

Pullups are added to counter the pulldowns, to allow the microcontroller to
simulate a detach (open-circuit) event.

Additionally, the microcontroller is upgraded to ATTINY48 (more pins) to allow
it to control these. This commit does not modify the firmware accordingly.

Signed-off-by: Chris Pavlina <pavlinac@ainfosec.com>